+++ app/src/main/java/com/daepiro/numberoneproject/presentation/view/funding/detail/SendHeartBottomSheet.kt
@@ -1,7 +1,6 @@
-package com.daepiro.numberoneproject.presentation.view.funding
+package com.daepiro.numberoneproject.presentation.view.funding.detail
 
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
