--- app/src/main/res/layout/item_funding_category.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <data>
-
-    </data>
-
-    <androidx.appcompat.widget.LinearLayoutCompat
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:orientation="vertical"
-        android:gravity="center_horizontal"
-        android:layout_marginRight="16dp">
-
-        <ImageView
-            android:id="@+id/iv_funding_category"
-            android:layout_width="52dp"
-            android:layout_height="52dp"
-            android:src="@drawable/ic_funding"
-            android:padding="12dp"
-            android:background="@drawable/bg_funding_category_select_on"/>
-
-        <TextView
-            android:id="@+id/tv_funding_category"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            style="@style/caption"
-            android:layout_marginTop="6dp"
-            android:text="@string/전체" />
-
-    </androidx.appcompat.widget.LinearLayoutCompat>
-</layout>
